pyproject.toml
[project]
name = "housing-prices"
version = "0.1.0"
description = "Economics-focused California housing price modeling pipeline"
authors = [{ name = "Sankalp Mishra" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas==2.2.2",
  "numpy==1.26.4",
  "scikit-learn==1.5.1",
  "matplotlib==3.9.0",
  "seaborn==0.13.2",
  "pyyaml==6.0.1",
  "plotly==5.22.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "black>=24.4",
  "ruff>=0.5.6"
]

[project.urls]
Homepage = "https://github.com/Sankalp232004/california-housing-project"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
minversion = "8.2"
addopts = "-ra"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
select = ["E", "F", "I"]
